backend/requirements.fastapi.txt
langchain_openai
streamlit
langchain_core
langchain
langchain_community
langchain_huggingface
python-dotenv
sentence-transformers
faiss-cpu
langgraph
uvicorn
pydantic
fastapi
rapidfuzz
